% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagecrop.R
\name{imagecrop}
\alias{imagecrop}
\title{Interface to the \code{imagemagic convert} tool for cropping images}
\usage{
imagecrop(fileName, outFileName = NULL)
}
\arguments{
\item{fileName}{character string denoting the path/ name of the input file}

\item{outFileName}{character string denoting the name of the output file. If
not specified, "_cr.[original file format]" is appended to \code{fileName}.}
}
\value{
Returns nothing, but as a side effect saves the cropped image file.
}
\description{
Automatically cropping of PNG, JPG, etc image files using the \code{convert}
comand line tool (\url{http://www.imagemagick.org/}) .
}
\examples{

library(networkD3)
data(MisLinks)
data(MisNodes)
p <- forceNetwork(Links = MisLinks, Nodes = MisNodes,
             Source = "source", Target = "target",
             Value = "value", NodeID = "name",
             Group = "group", opacity = 0.9,
             height = 900, width = 900)

# generated PNG needs a serious cropping
nwtbplots::jsGraphic2Png(p, c(900, 900, 20, 20), slow = TRUE)
imagecrop("JSgraphic.png", "JSgraphicCr1.png")

# generated PNG needs a serious cropping
nwtbplots::jsGraphic2Png(p, c(900, 900, 20, 20), slow = FALSE)
imagecrop("JSgraphic.png")

file.remove(c("JSgraphic.png", "JSgraphicCr1.png", "JSgraphic_cr.png"))

}
\seealso{
\pkg{knitr} \code{\link[knitr]{plot_crop}}, \pkg{ImageMagick}
  \url{http://www.imagemagick.org/}
}
\author{
Christoph Schmidt <schmidtchristoph@users.noreply.github.com>
}
