% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfPageExtract.R
\name{pdfPageExtract}
\alias{pdfPageExtract}
\title{Interface to \code{gs} to extract a range of pages from a PDF file}
\usage{
pdfPageExtract(fileName, firstP = NULL, lastP = NULL, outFileName = NULL)
}
\arguments{
\item{fileName}{character string denoting the path/ name of the input PDF
file}

\item{firstP}{integer denoting the first extracted page}

\item{lastP}{integer denoting the last extracted page}

\item{outFileName}{character string denoting the name of the output PDF file.
If not specified, it is
"[\code{fileName}]_pages[\code{firstP}]_[\code{lastP}].pdf".}
}
\value{
Returns nothing, but as a side effect saves the extracted pages as a
  new .pdf file.
}
\description{
Extract a range of pages from a PDF file. Preserves hyperlinks that fall
within the selected page range.
}
\note{
This function is based on Ghostscript.
}
\examples{

library(d3heatmap)
p <- d3heatmap(mtcars, scale = "column", colors = "Blues")
# pdf contains three pages (1 plot, 2 empty ones)
jsGraphic2Pdf(p, c(2500, 2000, 50, 50))
pdfPageExtract("JSgraphic.pdf", 1, 1)
pdfPageExtract("JSgraphic.pdf", 1, 1, "JSgraphic.pdf")

file.remove("JSgraphic.pdf", "JSgraphic_pages1_1.pdf", "JSgraphic_extr.pdf")

}
\seealso{
\pkg{GPL Ghostscript} \url{https://www.ghostscript.com}
}
\author{
Christoph Schmidt <schmidtchristoph@users.noreply.github.com>
}
